Rename python archives and apply POO
little changes to archive in english and apply a POO to do a modularization
